Crate polkadot_node_core_pvf_worker

Macros

decl_puppet_worker_main

Use this macro to declare a fn main() {} that will check the arguments and dispatch them to the appropriate worker, making the executable that can be used for spawning workers.

decl_worker_main

Use this macro to declare a fn main() {} that will create an executable that can be used for spawning the desired worker.

Functions

execute_worker_entrypoint

The entrypoint that the spawned execute worker should start with. The socket_path specifies the path to the socket used to communicate with the host. The node_version, if Some, is checked against the worker version. A mismatch results in immediate worker termination. None is used for tests and in other situations when version check is not necessary.

prepare

Runs preparation on the given runtime blob. If successful, it returns a serialized compiled artifact which can then be used to pass into Executor::execute after writing it to the disk.

prepare_worker_entrypoint

The entrypoint that the spawned prepare worker should start with. The socket_path specifies the path to the socket used to communicate with the host. The node_version, if Some, is checked against the worker version. A mismatch results in immediate worker termination. None is used for tests and in other situations when version check is not necessary.

prevalidate

Runs the prevalidation on the given code. Returns a [RuntimeBlob] if it succeeds.